Scan for boot files at run time.
This patch allows usbbootgui to dynamically load boot files at run
time. XDG standard directories are searched. Usually this means:

    /usr/share/rpiboot
    /usr/local/share/rpiboot
    ~/.local/share/rpiboot

The code first checks if the rpiboot/ subdirectory exists. If it
does, every subdirectory is checked for a bootcode.bin file. If
found, the directory is added to the list store.

If a description.txt is found in the directory, it is used as the
"friendly name" of the image. If an icon.png is found, it is loaded
as the icon. If not found, the directory name and a default icon
are used instead.

Both the description and path of the image are displayed in the GUI.
This is to distinguish between the same image in /usr/share and
/usr/local/share for example.

Images are now always refered to by their full paths. I have not
checked how this interacts with the "always use image" code, but I
think it should work. Old config files might have to be deleted.

This patch also allows the TreeView to have a scrollbar, and fixes
the "expand" property on a few items in the GUI, so that only the
TreeView expands when the window is resized.

Signed-off-by: Alistair Buxton <a.j.buxton@gmail.com>